add a required option -userid
Script will not function at all if -userid is not specified. Causes the following line in output:

  Use of uninitialized value $st_user in concatenation (.) or string at ./apg_run_script.pl line 499.

Traced this back to function call st_db_get_userid($userid) on line 198, which depends on line 63 and has no default value. Specifying this on the command line allows the script to execute past this point with no such error message in the output.