bug fix: uninitialized var $l_rule_uid
The 'if' statement is processed in order from left to right, so if no rule is found it tries to check the uninitialized variable $l_rule_uid before checking the flag $l_found_rule. It makes more sense to check the flag first before checking the contents of a variable which might not have been set to anything under certain conditions (such as when no rule UUID was found). Rearranging the order of the 'if' statement avoids this unnecessary error; if the flag isn't set, it doesn't evaluate the second condition at all.