replace 'order' field with 'rule_number' field
It seems the 'rule_number' field is the correct field to use to match on rule number rather than the 'order' field, at least for for Check Point devices. I am not certain that this is the case for Netscreen, Fortinet or Cisco devices, but I think it's safe to assume the database design is the same for all vendors. 

This fixes issue #9 as far as I can tell.